feat(project): 프로젝트 공고 회사명/클라이언트명 표기 추가
- 프로젝트 공고마다 회사명 혹은 클라이언트 명을 표기
- 개인 클라이언트 유저의 경우 회사명이 null값이기 때문에 User Name을 표기

+++ src/main/java/com/sanaiclub/project/dao/ProjectDashboardMapper.java
@@ -34,5 +34,5 @@ public interface ProjectDashboardMapper {
 
     int deadlineWithin7Days();
 
-    ProjectsVO selectProjectDetail(long projectId);
+    ProjectsVO selectProjectDetail(Integer projectId);
 }
